Filter out mate-unmapped for arcs
