+++ plugins/alignments/src/PileupRPC/methods/GetReducedFeatures.ts
@@ -1,12 +1,11 @@
 import { getAdapter } from '@jbrowse/core/data_adapters/dataAdapterCache'
-import { Region, dedupe } from '@jbrowse/core/util'
+import { Region, dedupe, groupBy } from '@jbrowse/core/util'
 import { RemoteAbortSignal } from '@jbrowse/core/rpc/remoteAbortSignals'
 import { BaseFeatureDataAdapter } from '@jbrowse/core/data_adapters/BaseAdapter'
 import { toArray } from 'rxjs/operators'
 import { firstValueFrom } from 'rxjs'
 // locals
 import { filterForPairs, getInsertSizeStats } from '../util'
-import { ReducedFeature } from '../../shared/fetchChains'
 import PileupBaseRPC from '../base'
 import { getTag } from '../../util'
 
@@ -56,15 +55,8 @@ export default class PileupGetReducedFeatures extends PileupBaseRPC {
 
     const filtered = filterForPairs(reduced)
     const stats = filtered.length ? getInsertSizeStats(filtered) : undefined
-    const chains = {} as { [key: string]: ReducedFeature[] }
+    const chains = groupBy(reduced, f => f.name)
 
-    // pair features
-    reduced.forEach(f => {
-      if (!chains[f.name]) {
-        chains[f.name] = []
-      }
-      chains[f.name].push(f)
-    })
     return {
       chains: Object.values(chains),
       stats,
