+++ plugins/wiggle/src/MultiRowLineRenderer/MultiRowLineRenderer.ts
@@ -1,5 +1,4 @@
-import { Feature } from '@jbrowse/core/util'
-import { groupBy } from '../util'
+import { groupBy, Feature } from '@jbrowse/core/util'
 import { drawLine } from '../drawxy'
 import WiggleBaseRenderer, {
   MultiRenderArgsDeserialized as MultiArgs,
