+++ plugins/wiggle/src/util.ts
@@ -167,15 +167,6 @@ export function getNiceDomain({
   return scale.domain() as [number, number]
 }
 
-export function groupBy<T>(array: T[], predicate: (v: T) => string) {
-  const result = {} as { [key: string]: T[] }
-  for (const value of array) {
-    const entry = (result[predicate(value)] ||= [])
-    entry.push(value)
-  }
-  return result
-}
-
 export async function getQuantitativeStats(
   self: {
     adapterConfig: AnyConfigurationModel
